compile llvm bitcode to assembly instead of 'C' code.
git-svn-id: http://llvm-lua.googlecode.com/svn/trunk@49 f87b780f-1855-0410-a4d3-fd11cda3f415

+++ llvm-lua/compile_bc.sh
@@ -1,15 +1,21 @@
 #!/bin/sh
 #
+ARCH=i686
+#ARCH=pentium4
+#ARCH=athlon64
 FILE=${1/.lua/}
 DEBUG="0"
 if [[ $DEBUG == "1" ]]; then
+	LUA_FLAGS=" -O3 "
+	#LUA_FLAGS=" -O3 -do-not-inline-opcodes -dump-functions "
 	CFLAGS=" -O0 -ggdb "
 	OPT_FLAGS=" -disable-opt "
 	LLC_FLAGS=" "
 else
-	CFLAGS=" -march=athlon64 -O3 -fomit-frame-pointer -pipe -Wall "
+	LUA_FLAGS=" -O3 "
+	CFLAGS=" -march=$ARCH -O3 -fomit-frame-pointer -pipe -Wall "
+	#CFLAGS=" -O3 -fomit-frame-pointer -pipe -Wall "
 	OPT_FLAGS=" -O3 -std-compile-opts -tailcallelim -tailduplicate "
-	#CFLAGS=" -march=athlon64 -O2 -pipe -Wall "
 	#OPT_FLAGS=" -disable-opt -tailcallelim "
 	LLC_FLAGS=" -tailcallopt "
 fi
@@ -20,7 +26,9 @@ fi
 	llc $LLC_FLAGS -f -filetype=asm -o ${FILE}_run.s ${FILE}_opt.bc && \
 	g++ $CFLAGS -o ${FILE} ${FILE}_run.s -lm -ldl
 
-rm -f ${FILE}_opt.bc ${FILE}_run.bc
-rm -f ${FILE}.bc ${FILE}_run.s
-#mv ${FILE}_opt.bc ${FILE}_run.bc
+if [[ $DEBUG == "0" ]]; then
+	rm -f ${FILE}_opt.bc ${FILE}_run.bc
+	rm -f ${FILE}.bc ${FILE}_run.s
+	#mv ${FILE}_opt.bc ${FILE}_run.bc
+fi
